requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
Pillow==8.1.2
pytz==2020.1
sqlparse==0.3.1
